Extracting model versioning code to a concern
Preparing to add versioning to forms. This pulls the common code from
ResponseSet and Question into an ActiveSupport::Concern.
