+++ app/models/response_set.rb
@@ -1,4 +1,6 @@
 class ResponseSet < ApplicationRecord
+  include Versionable
+
   has_many :question_response_sets
   has_many :questions, through: :question_response_sets
   has_many :responses, dependent: :nullify
@@ -8,39 +10,6 @@ class ResponseSet < ApplicationRecord
   belongs_to :updated_by, class_name: 'User'
   belongs_to :parent, class_name: 'ResponseSet'
   accepts_nested_attributes_for :responses, allow_destroy: true
-
-  validates :version_independent_id, presence: true,
-                                     if: proc { |rs| rs.version > 1 }
-  validates :version, presence: true, uniqueness: { scope: :version_independent_id,
-                                                    message: 'versions should be unique across a revised response set' }
-
-  after_save :assign_version_independent_id,
-             if: proc { |rs| rs.version == 1 && rs.version_independent_id.blank? }
-
-  def self.search(search)
-    where('name ILIKE ?', "%#{search}%")
-  end
-
-  # Finds only the latest versions of ResponseSets. ResponseSet.all will return
-  # all versions of all ResponseSets.
-  def self.latest_versions
-    joins('INNER JOIN (SELECT version_independent_id, MAX(version) as version
-             FROM response_sets GROUP BY version_independent_id) rs
-             ON rs.version_independent_id = response_sets.version_independent_id
-             AND response_sets.version = rs.version')
-  end
-
-  # Callback that assigns the version_independent_id. This should never be called
-  # directly.
-  def assign_version_independent_id
-    if version == 1
-      update_attribute(:version_independent_id, "RS-#{id}")
-    else
-      # If you have a ResponseSet with a version greater than 1, it should
-      # already have a version_independent_id set.
-      raise ActiveRecord::RecordInvalid, self
-    end
-  end
 
   # Builds a new ResponseSet object with the same version_independent_id. Increments
   # the version by one and builds a new set of Response objects to go with it.
@@ -53,38 +22,6 @@ class ResponseSet < ApplicationRecord
     end
 
     new_revision
-  end
-
-  # Finds any other versions of this ResponseSet in descending version order
-  def other_versions
-    ResponseSet.where(version_independent_id: version_independent_id)
-               .where.not(version: version)
-               .order(version: :desc)
-  end
-
-  def all_versions
-    ResponseSet.where(version_independent_id: version_independent_id)
-               .order(version: :desc)
-  end
-
-  def most_recent
-    if other_versions.present?
-      if other_versions[0].version > version
-        other_versions[0].version
-      else
-        version
-      end
-    end
-  end
-
-  def most_recent?
-    if other_versions.present?
-      if other_versions[0].version > version
-        false
-      else
-        true
-      end
-    end
   end
 
   def extend_from
